Add JOB-FAILED condition and signal it on WORK_FAIL response.
This change allows to retry job only if the job was failed.

Now client method SUBMIT-JOB signals a special error JOB-FAILED. Previously
it used standard ERROR.

Also, it uses restarts RETRY-JOB and ACCEPT-JOB, used in the worker code as well.
Previously, it used RETRY and ACCEPT names for these retries.